Move NeovimInstance to eovim folder
