+++ browser/src/Services/CommandManager.ts
@@ -6,7 +6,7 @@
 
 import * as _ from "lodash"
 
-import { INeovimInstance } from "./../NeovimInstance"
+import { INeovimInstance } from "./../neovim"
 
 import { ITask, ITaskProvider } from "./Tasks"
