+++ browser/src/Services/Errors.ts
@@ -1,6 +1,6 @@
 import * as _ from "lodash"
 
-import { INeovimInstance } from "./../NeovimInstance"
+import { INeovimInstance } from "./../neovim"
 import * as Performance from "./../Performance"
 
 import { ITask, ITaskProvider } from "./Tasks"
