+++ browser/src/neovim/QuickFix.ts
@@ -1,4 +1,4 @@
-import { INeovimInstance } from "../NeovimInstance"
+import { INeovimInstance } from "./NeovimInstance"
 
 export type Action = "a" | " "
