+++ browser/src/neovim/index.ts
@@ -1,3 +1,6 @@
+export * from "./Buffer"
 export * from "./MsgPack"
 export * from "./Session"
 export * from "./NeovimProcessSpawner"
+export * from "./NeovimInstance"
+export * from "./Window"
